requirements.txt
Django	3.1	
Faker	10.0.0	
Jinja2	2.11.2	
MarkupSafe	1.1.1	
Pillow	7.2.0	
PyJWT	1.7.1	
asgiref	3.2.10	
certifi	2020.6.20	
cffi	1.14.3	
chardet	3.0.4	
coreapi	2.3.3	
coreschema	0.0.4	
cryptography	3.1.1	
defusedxml	0.6.0	
django-admin-relation-links	0.2.5	
django-allauth	0.42.0	
django-braces	1.14.0	
django-ckeditor	6.0.0	
django-filter	2.4.0	
django-js-asset	1.2.2	
django-modeltranslation	0.15.2	
django-oauth-toolkit	1.3.2	
django-rest-framework-social-oauth2	1.1.0	
django-templated-mail	1.1.1	
djangorestframework	3.12.1	
djangorestframework-simplejwt	4.4.0	
djoser	2.0.5	
drf-yasg	1.17.1	
idna	2.10	
inflection	0.5.1	
itypes	1.2.0	
oauthlib	3.1.0	
packaging	20.4	
pip	21.1.1	
psycopg2-binary	2.8.6	
pycparser	2.20	
pyparsing	2.4.7	
python-dateutil	2.8.2	
python3-openid	3.2.0	
pytz	2020.1	
requests	2.24.0	
requests-oauthlib	1.3.0	
ruamel.yaml	0.16.12	
ruamel.yaml.clib	0.2.2	
setuptools	49.4.0	
six	1.15.0	
social-auth-app-django	4.0.0	
social-auth-core	3.3.3	
sqlparse	0.3.1	
text-unidecode	1.3	
uritemplate	3.0.1	
urllib3	1.25.10	
wheel	0.35.1	
